+++ plugins/decl_netcdf/tests/decl_netcdf_tests.cxx
@@ -23,8 +23,11 @@
  * THE SOFTWARE.
  ******************************************************************************/
 
-#include <gtest/gtest.h>
-#include <pdi.h>
+#include <filesystem>
+#include <numeric>
+#include <ranges>
+
+#include <pdi/testing.h>
 
 /*
  * Name:                decl_netcdf_test.01
@@ -1023,90 +1026,269 @@ TEST(decl_netcdf_test, deflate)
 	PDI_finalize();
 }
 
+class DeclNetcdfCheckType: public ::PDI::PdiTest
+{};
+
 /*
- * Name:                decl_netcdf_test.read
+ * Name:                DeclNetcdfCheckType.IntReadMismatch
  *
  * Description:         Tests write and read of int with type mismatch
  */
-TEST(decl_netcdf_test, int_read)
+TEST_F(DeclNetcdfCheckType, IntReadMismatch)
 {
-	constexpr char CONFIG_YAML[] = R"(
-    logging: trace
-    data:
-      int_in: int32
-      int_out: int64
-    plugins:
-      decl_netcdf:
-      - file: 'test_int_read.nc'
-        variable:
-          scalar_int32: int32
-        on_event: write_data
-        write:
-          int_in:
-            variable: scalar_int32
-      - file: 'test_int_read.nc'
-        variable:
-          scalar_int32: int32
-        on_event: read_data
-        read:
-          int_out:
-            variable:
-              scalar_int32
-    )";
-
-	PDI_init(PC_parse_string(CONFIG_YAML));
-	PDI_errhandler(PDI_NULL_HANDLER);
+	InitPdi(PC_parse_string(R"==(
+logging: trace
+data:
+  int_in: int32
+  int_out: int64
+plugins:
+  decl_netcdf:
+    - file: 'test_int_read.nc'
+      on_event: write_data
+      write:
+        int_in:
+          variable: scalar_int32
+    - file: 'test_int_read.nc'
+      on_event: read_data
+      read:
+        int_out:
+          variable:
+            scalar_int32
+)=="));
 
 	// write data
 	int32_t int_in = 42;
 	PDI_multi_expose("write_data", "int_in", &int_in, PDI_OUT, NULL);
 
+	EXPECT_TRUE(std::filesystem::exists("test_int_read.nc"));
+
 	// read data
 	int64_t int_out = -1;
-	EXPECT_NE(PDI_OK, PDI_multi_expose("read_data", "int_out", &int_out, PDI_IN, NULL));
-	PDI_finalize();
+
+	EXPECT_CALL(
+		*this,
+		PdiError(
+			testing::Eq(PDI_ERR_TYPE),
+			testing::AllOf(
+				testing::HasSubstr("while triggering `read_data',"),
+				testing::HasSubstr("Decl_netcdf plugin: Datatype mismatch (with size): "
+	                               "read 'scalar_int32' of size 4 for a buffer of size 8")
+			)
+		)
+	);
+
+	EXPECT_EQ(PDI_ERR_TYPE, PDI_multi_expose("read_data", "int_out", &int_out, PDI_IN, NULL));
 }
 
 /*
- * Name:                decl_netcdf_test.read
+ * Name:                DeclNetcdfCheckType.FloatReadMismatch
  *
  * Description:         Tests write and read of float/double with type mismatch
  */
-TEST(decl_netcdf_test, float_read)
+TEST_F(DeclNetcdfCheckType, FloatReadMismatch)
 {
-	constexpr char CONFIG_YAML[] = R"(
-    logging: trace
-    data:
-      var_in: float
-      var_out: double
-    plugins:
-      decl_netcdf:
-      - file: 'test_float_read.nc'
-        variable:
-          nc_var: float
-        on_event: write_data
-        write:
-          var_in:
-            variable: nc_var
-      - file: 'test_float_read.nc'
-        variable:
-          nc_var: float
-        on_event: read_data
-        read:
-          var_out:
-            variable: nc_var
-    )";
-
-	PDI_init(PC_parse_string(CONFIG_YAML));
-	PDI_errhandler(PDI_NULL_HANDLER);
+	InitPdi(PC_parse_string(R"==(
+logging: trace
+data:
+  var_in: float
+  var_out: double
+plugins:
+  decl_netcdf:
+    - file: 'test_float_read.nc'
+      on_event: write_data
+      write:
+        var_in:
+          variable: scalar_float
+    - file: 'test_float_read.nc'
+      on_event: read_data
+      read:
+        var_out:
+          variable: scalar_float
+)=="));
 
 	// write data
 	float var_in = 12.34;
 	PDI_multi_expose("write_data", "var_in", &var_in, PDI_OUT, NULL);
 
+	EXPECT_TRUE(std::filesystem::exists("test_float_read.nc"));
+
+	EXPECT_CALL(
+		*this,
+		PdiError(
+			testing::Eq(PDI_ERR_TYPE),
+			testing::AllOf(
+				testing::HasSubstr("while triggering `read_data',"),
+				testing::HasSubstr("Decl_netcdf plugin: Datatype mismatch (with size): "
+	                               "read 'scalar_float' of size 4 for a buffer of size 8")
+			)
+		)
+	);
+
 	// read data
 	double var_out = -1.0;
-	EXPECT_NE(PDI_OK, PDI_multi_expose("read_data", "var_out", &var_out, PDI_IN, NULL));
+	EXPECT_EQ(PDI_ERR_TYPE, PDI_multi_expose("read_data", "var_out", &var_out, PDI_IN, NULL));
+}
 
-	PDI_finalize();
+/*
+ * Name:                DeclNetcdfCheckType.ReadDataNotDefinedInYaml
+ *
+ * Description:         Tests write and read of float/double which is not defined in Yaml
+ *                      Read with on_event
+ */
+TEST_F(DeclNetcdfCheckType, ReadDataNotDefinedInYamlCaseOnEvent)
+{
+	InitPdi(PC_parse_string(R"==(
+logging: trace
+data:
+  var_in: float
+plugins:
+  decl_netcdf:
+    - file: 'test_float_read_data_not_defined.nc'
+      on_event: write_data
+      write:
+        var_in:
+          variable: scalar_float
+    - file: 'test_float_read_data_not_defined.nc'
+      on_event: read_data
+      read:
+        var_out:
+          variable: scalar_float
+)=="));
+
+	// write data
+	float var_in = 15.34;
+	PDI_multi_expose("write_data", "var_in", &var_in, PDI_OUT, NULL);
+
+	EXPECT_TRUE(std::filesystem::exists("test_float_read_data_not_defined.nc"));
+
+	EXPECT_CALL(
+		*this,
+		PdiError(
+			testing::Eq(PDI_ERR_TYPE),
+			testing::AllOf(
+				testing::HasSubstr("while triggering `read_data',"),
+				testing::HasSubstr("Can not read `scalar_float' :"),
+				testing::HasSubstr("Invalid type in Decl_netcdf plugin:"),
+				testing::HasSubstr("The exposed data `var_out' "
+	                               "is not defined in yaml (meta)data section.")
+			)
+		)
+	);
+
+	// read data
+	float var_out = -1.0;
+	EXPECT_EQ(PDI_ERR_TYPE, PDI_multi_expose("read_data", "var_out", &var_out, PDI_IN, NULL));
+}
+
+/*
+ * Name:                DeclNetcdfCheckType.ReadDataNotDefinedInYamlCaseOnData
+ *
+ * Description:         Tests write and read of float/double which is not defined in Yaml
+ *                      Read with on_data
+ */
+TEST_F(DeclNetcdfCheckType, ReadDataNotDefinedInYamlCaseOnData)
+{
+	InitPdi(PC_parse_string(R"==(
+logging: trace
+data:
+  var_in: float
+plugins:
+  decl_netcdf:
+    - file: 'test_float_read_data_not_defined.nc'
+      on_event: write_data
+      write:
+        var_in:
+          variable: scalar_float
+    - file: 'test_float_read_data_not_defined.nc'
+      on_data: var_out
+      read:
+        var_out:
+          variable: scalar_float
+)=="));
+
+	// write data
+	float var_in = 15.34;
+	PDI_multi_expose("write_data", "var_in", &var_in, PDI_OUT, NULL);
+
+	EXPECT_TRUE(std::filesystem::exists("test_float_read_data_not_defined.nc"));
+
+	EXPECT_CALL(
+		*this,
+		PdiError(
+			testing::Eq(PDI_ERR_TYPE),
+			testing::AllOf(
+				testing::HasSubstr("while sharing `var_out'"),
+				testing::HasSubstr("Can not read `scalar_float'"),
+				testing::HasSubstr("Invalid type in Decl_netcdf plugin:"),
+				testing::HasSubstr("The exposed data `var_out' "
+	                               "is not defined in yaml (meta)data section.")
+			)
+		)
+	);
+
+	// read data
+	float var_out = -1.0;
+	PDI_expose("var_out", &var_out, PDI_IN);
+}
+
+/*
+ * Name:                DeclNetcdfCheckType.ReadDoubleArrayNotDefinedInYaml
+ *
+ * Description:         Tests write and read of double array that not defined in Yaml
+ */
+TEST_F(DeclNetcdfCheckType, ReadDoubleArrayNotDefinedInYaml)
+{
+	InitPdi(PC_parse_string(R"==(
+logging: trace
+metadata:
+  nn: int
+data:
+  array_in: {type: array, subtype: double, size: ['$nn']}
+plugins:
+  decl_netcdf:
+    - file: 'test_double_array_read.nc'
+      variables:
+        nc_var: {type: array, subtype: double, size: ['$nn']}
+      on_event: write_data
+      write:
+        array_in:
+          variable: nc_var
+    - file: 'test_double_array_read.nc'
+      variables:
+        nc_var:  {type: array, subtype: double, size: ['$nn']}
+      on_event: read_data
+      read:
+        array_out:
+          variable: nc_var
+)=="));
+
+
+	int const nn = 3;
+	PDI_expose("nn", &nn, PDI_INOUT);
+
+	auto const array_in = make_a<std::array<double, nn>>();
+
+	PDI_multi_expose("write_data", "array_in", array_in.data(), PDI_OUT, NULL);
+
+	// read data
+	std::array<double, nn> array_out{};
+	array_out.fill(-1.0);
+
+	EXPECT_TRUE(std::filesystem::exists("test_double_array_read.nc"));
+
+	EXPECT_CALL(
+		*this,
+		PdiError(
+			testing::Eq(PDI_ERR_TYPE),
+			testing::AllOf(
+				testing::HasSubstr("while triggering `read_data',"),
+				testing::HasSubstr("Can not read `nc_var'"),
+				testing::HasSubstr("Invalid type in Decl_netcdf plugin:"),
+				testing::HasSubstr("The exposed data `array_out' "
+	                               "is not defined in yaml (meta)data section.")
+			)
+		)
+	);
+
+	PDI_multi_expose("read_data", "array_out", array_out.data(), PDI_IN, NULL);
 }