+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to
 ### For users
 
 #### Added
+* Improved error message in type check in Decl'NetCDF when reading scalar variable from file
+[#731]https://github.com/pdidev/pdi/issues/731
 
 #### Changed
 
@@ -97,7 +99,6 @@ and this project adheres to
 
 #### Fixed
 * Fixed the build failure of a test in 1.11.1
-
 
 
 ## [1.11.1] - 2026-06-23